Extract TrailHistoryExport.Export
